requirements.txt
# Core agent framework
langgraph>=0.2.0
langchain-google-genai>=2.0.0
google-genai>=1.0.0

# API layer
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# Sandbox execution
e2b-code-interpreter>=1.0.0

# AI coding assistant
aider-chat>=0.50.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
